FloatingPoint: Minor bug fixes.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@45828 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/floatingpoint/MyFloat.java
@@ -665,10 +665,10 @@ public class MyFloat {
 
   public double toDouble() {
     if (isNan()) {
-      return Float.NaN;
+      return Double.NaN;
     }
     if (isInfinite()) {
-      return isNegative() ? Float.NEGATIVE_INFINITY : Float.POSITIVE_INFINITY;
+      return isNegative() ? Double.NEGATIVE_INFINITY : Double.POSITIVE_INFINITY;
     }
     return new BigFloat(value.sign, value.significand, value.exponent, BinaryMathContext.BINARY64)
         .doubleValueExact();
